# backend/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plan2Field AI",
    "version": "0.1.0",
    "description": "Local prototype. All actual mutations require project review. No external scheduling writes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/projects/{projectId}/reports": {
      "get": {
        "summary": "List project reports",
        "operationId": "get__projects_projectId_reports",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      },
      "post": {
        "summary": "Queue idempotent report processing",
        "operationId": "post__projects_projectId_reports",
        "responses": {
          "202": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token from /auth/me. Browser Origin must equal APP_ORIGIN."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ReportSubmission"
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/reports/{reportId}": {
      "get": {
        "summary": "Get scoped report",
        "operationId": "get__projects_projectId_reports_reportId",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/api/projects/{projectId}/reports/{reportId}/audio": {
      "get": {
        "summary": "Authorized original audio replay",
        "operationId": "get__projects_projectId_reports_reportId_audio",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/projects/{projectId}/reports/{reportId}/clarifications": {
      "post": {
        "summary": "Append clarification as new report evidence",
        "operationId": "post__projects_projectId_reports_reportId_clarifications",
        "responses": {
          "202": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token from /auth/me. Browser Origin must equal APP_ORIGIN."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSubmission"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "ReportSubmission": {
        "type": "object",
        "required": [
          "source_type",
          "reporting_date",
          "captured_at",
          "idempotency_key"
        ],
        "properties": {
          "source_type": {
            "enum": [
              "TEXT",
              "VOICE",
              "SPREADSHEET",
              "SCAN"
            ]
          },
          "original_text": {
            "type": "string",
            "maxLength": 50000
          },
          "reporting_date": {
            "type": "string",
            "format": "date"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time"
          },
          "idempotency_key": {
            "type": "string",
            "format": "uuid"
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "mapping": {
            "type": "string",
            "description": "JSON object mapping text to a spreadsheet column header"
          }
        }
      }
    }
  }
}